# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Immobilière",
        "description": "API pour la gestion des biens immobiliers",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tenant/contact/agency": {
            "post": {
                "tags": [
                    "Locataire"
                ],
                "summary": "Envoyer un message à l'agence",
                "description": "Permet au locataire authentifié d'envoyer un email à son agence.",
                "operationId": "sendEmailToAgency",
                "requestBody": {
                    "description": "Données du message à envoyer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subject",
                                    "message",
                                    "agency_email"
                                ],
                                "properties": {
                                    "subject": {
                                        "type": "string",
                                        "example": "Problème avec la plomberie"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Bonjour, je vous contacte car j'ai une fuite d'eau dans la salle de bain."
                                    },
                                    "agency_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "agence.immobiliere@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message envoyé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentification requise"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur serveur: ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}